watchのコールバック関数を別ファイルに分離
